requirements.txt
absl-py==1.2.0
aiohttp==3.8.1
aiosignal==1.2.0
altair==4.2.0
astunparse==1.6.3
async-timeout==4.0.2
attrs==22.1.0
base58==2.1.1
bitarray==2.6.0
blinker==1.5
cachetools==5.2.0
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
cmake==3.24.2
colorama==0.4.5
commonmark==0.9.1
cryptography==38.0.1
cycler==0.11.0
cytoolz==0.12.0
DateTime==4.7
decorator==5.1.1
entrypoints==0.4
eth-abi==3.0.1
eth-account==0.7.0
eth-hash==0.3.3
eth-keyfile==0.6.0
eth-keys==0.4.0
eth-rlp==0.3.0
eth-tester==0.7.0b1
eth-typing==3.2.0
eth-utils==2.0.0
face-recognition-models==0.3.0
fer==22.4.0
flatbuffers==1.12
fonttools==4.37.1
frozenlist==1.3.1
gast==0.4.0
gitdb==4.0.9
GitPython==3.1.27
google-auth==2.11.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.48.1
h5py==3.7.0
hexbytes==0.3.0
idna==3.3
importlib-metadata==4.12.0
ipfshttpclient==0.8.0a2
Jinja2==3.1.2
joblib==1.1.0
jsonschema==4.15.0
keras==2.9.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.4
libclang==14.0.6
lru-dict==1.1.8
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib==3.5.3
mtcnn==0.1.1
multiaddr==0.0.9
multidict==6.0.2
netaddr==0.8.0
numpy==1.23.2
oauthlib==3.2.0
opencv-contrib-python==4.6.0.66
opencv-python==4.6.0.66
opt-einsum==3.3.0
packaging==21.3
pandas==1.4.4
parsimonious==0.8.1
pdfminer.six==20220524
pdfplumber==0.7.4
pickle-mixin==1.0.2
Pillow==9.2.0
protobuf==3.19.4
py-solc-x==1.1.1
pyarrow==9.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pycryptodome==3.15.0
pydeck==0.8.0b1
Pygments==2.13.0
Pympler==1.0.1
pyparsing==3.0.9
PyPDF2==2.11.1
pyrsistent==0.18.1
python-dateutil==2.8.2
pytz==2022.2.1
pytz-deprecation-shim==0.1.0.post0
pywin32==304
requests==2.28.1
requests-oauthlib==1.3.1
rich==12.5.1
rlp==3.0.0
rsa==4.9
scikit-learn==1.1.2
scipy==1.9.1
semantic-version==2.10.0
semver==2.13.0
six==1.16.0
sklearn==0.0
smmap==5.0.0
streamlit==1.12.2
streamlit-option-menu==0.3.2
tensorboard==2.9.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.9.1
tensorflow-estimator==2.9.0
tensorflow-io-gcs-filesystem==0.26.0
termcolor==1.1.0
threadpoolctl==3.1.0
toml==0.10.2
toolz==0.12.0
tornado==6.2
tqdm==4.64.0
typing_extensions==4.3.0
tzdata==2022.2
tzlocal==4.2
urllib3==1.26.12
validators==0.20.0
varint==1.0.2
Wand==0.6.10
watchdog==2.1.9
web3==5.30.0
websockets==9.1
Werkzeug==2.2.2
wrapt==1.14.1
yarl==1.8.1
zipp==3.8.1
zope.interface==5.5.1
